+++ examples/using-javascript-transforms/src/components/Layouts/blogPost.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { Link } from "gatsby"
+import { Link, graphql } from "gatsby"
 
 import MasterLayout from "./master"