+++ examples/using-mongodb/src/templates/item.js
@@ -1,4 +1,5 @@
 import React from "react"
+import { graphql } from "gatsby"
 import Layout from "../layouts"
 
 class Item extends React.Component {
@@ -28,7 +29,7 @@ class Item extends React.Component {
 export default Item
 
 export const pageQuery = graphql`
-  query ItemQuery($id: String!) {
+  query($id: String!) {
     mongodbCloudDocuments(id: { eq: $id }) {
       id
       name