+++ examples/using-sqip/src/layouts/index.js
@@ -43,7 +43,7 @@ Layout.propTypes = {
 export default Layout
 
 export const query = graphql`
-  query SiteTitleQuery {
+  query {
     images: allFile(
       filter: { sourceInstanceName: { eq: "images" }, ext: { eq: ".jpg" } }
     ) {