+++ packages/gatsby-plugin-netlify-cms/src/cms-identity.js
@@ -1,0 +1,4 @@
+import netlifyIdentityWidget from "netlify-identity-widget"
+
+window.netlifyIdentity = netlifyIdentityWidget
+netlifyIdentityWidget.init()