+++ packages/gatsby/src/internal-plugins/query-runner/query-runner.js
@@ -142,7 +142,7 @@ ${formatErrorDetails(errorDetails)}`)
       dataPath = `${modInt}/${dataPath}`
     }
 
-    await fs.writeFile(resultPath, resultJSON)
+    await fs.outputFile(resultPath, resultJSON)
 
     store.dispatch({
       type: `SET_JSON_DATA_PATH`,
@@ -152,6 +152,8 @@ ${formatErrorDetails(errorDetails)}`)
       },
     })
 
-    return
+    return result
   }
+
+  return result
 }