+++ packages/gatsby/src/redux/actions.js
@@ -1,6 +1,6 @@
 // @flow
-import Joi from "joi"
-import chalk from "chalk"
+const Joi = require(`joi`)
+const chalk = require(`chalk`)
 const _ = require(`lodash`)
 const { bindActionCreators } = require(`redux`)
 const { stripIndent } = require(`common-tags`)
@@ -12,8 +12,8 @@ const { hasNodeChanged, getNode } = require(`./index`)
 const { trackInlineObjectsInRootNode } = require(`../schema/node-tracking`)
 const { store } = require(`./index`)
 const fileExistsSync = require(`fs-exists-cached`).sync
-import * as joiSchemas from "../joi-schemas/joi"
-import { generateComponentChunkName } from "../utils/js-chunk-names"
+const joiSchemas = require(`../joi-schemas/joi`)
+const { generateComponentChunkName } = require(`../utils/js-chunk-names`)
 
 const actions = {}
 
@@ -54,6 +54,12 @@ type Page = {
 
 type Plugin = {
   name: string,
+}
+
+type ActionOptions = {
+  traceId: ?string,
+  parentSpan: ?Object,
+  followsSpan: ?Object,
 }
 
 /**
@@ -97,7 +103,11 @@ const hasWarnedForPageComponent = new Set()
  * })
  */
 const fileOkCache = {}
-actions.createPage = (page: PageInput, plugin?: Plugin, traceId?: string) => {
+actions.createPage = (
+  page: PageInput,
+  plugin?: Plugin,
+  actionOptions?: ActionOptions
+) => {
   let noPageOrComponent = false
   let name = `The plugin "${plugin.name}"`
   if (plugin.name === `default-site-plugin`) {
@@ -293,9 +303,9 @@ ${reservedFields.map(f => `  * "${f}"`).join(`\n`)}
   }
 
   return {
+    ...actionOptions,
     type: `CREATE_PAGE`,
     plugin,
-    traceId,
     payload: internalPage,
   }
 }
@@ -452,7 +462,11 @@ const typeOwners = {}
  *   }
  * })
  */
-actions.createNode = (node: any, plugin?: Plugin, traceId?: string) => {
+actions.createNode = (
+  node: any,
+  plugin?: Plugin,
+  actionOptions?: ActionOptions = {}
+) => {
   if (!_.isObject(node)) {
     return console.log(
       chalk.bold.red(
@@ -566,6 +580,11 @@ actions.createNode = (node: any, plugin?: Plugin, traceId?: string) => {
     }
   }
 
+  if (actionOptions.parentSpan) {
+    actionOptions.parentSpan.setTag(`nodeId`, node.id)
+    actionOptions.parentSpan.setTag(`nodeType`, node.id)
+  }
+
   let deleteAction
   let updateNodeAction
   // Check if the node has already been processed.
@@ -573,7 +592,7 @@ actions.createNode = (node: any, plugin?: Plugin, traceId?: string) => {
     updateNodeAction = {
       type: `TOUCH_NODE`,
       plugin,
-      traceId,
+      ...actionOptions,
       payload: node.id,
     }
   } else {
@@ -591,7 +610,7 @@ actions.createNode = (node: any, plugin?: Plugin, traceId?: string) => {
     updateNodeAction = {
       type: `CREATE_NODE`,
       plugin,
-      traceId,
+      ...actionOptions,
       payload: node,
     }
   }
@@ -669,7 +688,7 @@ type CreateNodeInput = {
 actions.createNodeField = (
   { node, name, value, fieldName, fieldValue }: CreateNodeInput,
   plugin: Plugin,
-  traceId?: string
+  actionOptions?: ActionOptions
 ) => {
   if (fieldName) {
     console.warn(
@@ -724,7 +743,7 @@ actions.createNodeField = (
   return {
     type: `ADD_FIELD_TO_NODE`,
     plugin,
-    traceId,
+    ...actionOptions,
     payload: node,
   }
 }